requirements.txt
pandas
scikit-learn
streamlit
requests
notebook
python-dotenv